Re-organize shard tracker utils
